% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_dataframe}
\alias{plot_dataframe}
\title{Plot all Variables in a Data Frame}
\usage{
plot_dataframe(df, makepdf = F, filename = "plot_dataframe")
}
\arguments{
\item{df}{The data.frame to be plotted}
}
\value{
None
}
\description{
Makes plots for all variables in a data.frame, each formatted as appropriate
Plots numeric variables as histograms, and factors as barplots of tables.
Expects long names and many values for factors, so plots them sideways,
instead of the default, which has names at the bottom of the plot.
}
\examples{
plot_dataframe(df)

}
