% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_monkeyfile}
\alias{read_monkeyfile}
\title{Read a Flying Monkey File}
\usage{
read_monkeyfile(df_original, filepath)
}
\arguments{
\item{df_original}{the data frame that is to be transformed}

\item{filepath}{the path to the variable spec file being read in}
}
\value{
A data.frame, transformed using the information in the monkeyfile
}
\description{
Reads a completed file, and uses it to transform the original data
}
\details{
For monkeyfile format, see ?write_monkeyfile
}
\examples{
read_monkeyfile(df, "Documents/specs.txt")

}
