% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sideways_plot}
\alias{sideways_plot}
\title{Plot one variable in an easy to read barchart}
\usage{
sideways_plot(x, y = NULL, header = "", color = "gray")
}
\arguments{
\item{x}{The data to be plotted}

\item{y}{An optional boolean vector}
}
\value{
None
}
\description{
With a single variable, makes a simple, cleaned-up, sideways-oriented
barchart ('barplot') to allow the value names to be read more easily.
With a variable and a boolean, plots overlaid distributions of
the variables that do/don't have a TRUE value, to allow for easy comparisons.
}
\examples{
sideways_plot(x=health, y=received_treatment)

}
