% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_monkeyfile}
\alias{write_monkeyfile}
\title{Write a Flying Monkey File}
\usage{
write_monkeyfile(df, filename)
}
\arguments{
\item{df}{the data frame that provides the variable information for the file}

\item{filename}{the name of the variable spec file being written}
}
\value{
None
}
\description{
Creates a partially pre-filled file with variable information for the user\cr
to edit
}
\details{
The format of the created monkeyfile will be:

OLD_NAME:\cr
(the old name of the variable, ex: "What is your name?")\cr
NEW_NAME:\cr
(created blank, filled in manually, ex: "name". If unused, defaults to old_name)\cr
TYPE:\cr
(the inferred class/type, ex: 'numeric', 'character', 'factor')\cr
OLD_VALUES: (only created if the inferred type is NOT string or numeric)\cr
value_1\cr
value_2\cr
value_n\cr
NEW_VALUES: (only created if the inferred type is NOT string or numeric)\cr
(left blank; if only reordering is required, you can re-order the OLD_VALUES.\cr
The element of highest value should go at the top, eg. Big; Medium; Small\cr
NEW_VALUES allows you to rename all of the elements in the variable -- just\cr
make sure their order matches that of the possibly re-ordered OLD_VALUES)\cr
- - - - - (five dashes denotes the end of a variable)\cr
}
\examples{
write_monkeyfile(df, "specs") # => specs.txt file to be edited manually

}
